[dev](const_var.py): added const var USE_STATIC_POSITION

+++ node-manager/main.py
@@ -274,16 +274,17 @@ def run(enable_load_awareness: bool, lofi_delta: float, lofi_n: int,
     # -------------------------------------------------------------------
     # start frr    added by sqsq
     start_frr(docker_client)
-    # # set the initial position
-    # for node_id in sorted(list(satellite_map.keys())):
-    #     satellite_node = satellite_map[node_id]
-    #     node_id_str = satellite_id_tuple_to_str(node_id)
-    #     position_datas[node_id_str][LATITUDE_KEY], \
-    #     position_datas[node_id_str][LONGITUDE_KEY], \
-    #     position_datas[node_id_str][HEIGHT_KEY] = satellite_node.get_next_position(TIME_BASE)
-    #     # logger.info(f"{node_id_str}: {position_datas[node_id_str]}")
-    # update_network_delay(docker_client, position_datas, connect_order_map)
-    # time.sleep(WARMUP_PERIOD)
+    if USE_STATIC_POSITION:
+        # set the initial position
+        for node_id in sorted(list(satellite_map.keys())):
+            satellite_node = satellite_map[node_id]
+            node_id_str = satellite_id_tuple_to_str(node_id)
+            position_datas[node_id_str][LATITUDE_KEY], \
+            position_datas[node_id_str][LONGITUDE_KEY], \
+            position_datas[node_id_str][HEIGHT_KEY] = satellite_node.get_next_position(TIME_BASE)
+            # logger.info(f"{node_id_str}: {position_datas[node_id_str]}")
+        update_network_delay(docker_client, position_datas, connect_order_map)
+    time.sleep(WARMUP_PERIOD)
     # -------------------------------------------------------------------
     # -------------------------------------------------------------------
         
@@ -316,7 +317,8 @@ def run(enable_load_awareness: bool, lofi_delta: float, lofi_n: int,
                                                                          updater,
                                                                          BROADCAST_SEND_INTERVAL,
                                                                          connect_order_map))
-    update_position_process.start()
+    if not USE_STATIC_POSITION:
+        update_position_process.start()
     # ----------------------------------------------------------
 
     # start link failure generation and UDP send & recv
@@ -334,7 +336,8 @@ def run(enable_load_awareness: bool, lofi_delta: float, lofi_n: int,
     # ----------------------------------------------------------
     # clear after one run
     set_monitor_process.kill()
-    update_position_process.kill()
+    if not USE_STATIC_POSITION:
+        update_position_process.kill()
     delete_constellation(docker_client)
 
     uninstall_kernel_modules()
